test: cover unset TBO CPU token count
Cover eager and CUDA graph capture batches where
num_token_non_padded_cpu is unset and verify the
fallback to physical input rows.
